+++ src/print/node-helpers.ts
@@ -20,6 +20,7 @@ import {
     ThenBlockNode,
     CommentNode,
     SlotTemplateNode,
+    StyleDirectiveNode,
 } from './nodes';
 import { blockElements, TagName } from '../lib/elements';
 import { FastPath, ParserOptions } from 'prettier';
@@ -281,7 +282,7 @@ export function isAttributeShorthand(node: true | Node[]): node is [AttributeSho
 /**
  * True if node is of type `{a}` or `a={a}`
  */
-export function isOrCanBeConvertedToShorthand(node: AttributeNode): boolean {
+export function isOrCanBeConvertedToShorthand(node: AttributeNode | StyleDirectiveNode): boolean {
     if (isAttributeShorthand(node.value)) {
         return true;
     }
@@ -395,6 +396,10 @@ export function shouldHugStart(
         return true;
     }
 
+    if (options.htmlWhitespaceSensitivity === 'ignore') {
+        return false;
+    }
+
     const firstChild = children[0];
     return !isTextNodeStartingWithWhitespace(firstChild);
 }
@@ -423,6 +428,10 @@ export function shouldHugEnd(
     const children: Node[] = node.children;
     if (children.length === 0) {
         return true;
+    }
+
+    if (options.htmlWhitespaceSensitivity === 'ignore') {
+        return false;
     }
 
     const lastChild = children[children.length - 1];